Separate rbac middleware into own file
